backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.2
djangorestframework-simplejwt==4.8.0
psycopg2-binary==2.9.3
isort>=5.12.0
drf-base64>=2.0
python-dotenv==1.0.0
Pillow==11.0.0
djoser==2.1.0
gunicorn==20.1.0
black==23.7.0
flake8==6.0.0
flake8-isort==6.0.0
reportlab>=4.0.4
sqlparse>=0.4.4
pep8-naming==0.13.3
